rename files, move Enum to own file
