+++ WebhookScripts/Receiver/src/enums.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class AlertType(str, Enum):
+    MOTION_ALERT = 'motion_alert'
